redux-toolkit설치 후, slice생성/action,reducer파일 삭제/ redux-thunk삭제
