+++ package.json
@@ -41,11 +41,11 @@
     "webpack-dev-server": "^4.8.1"
   },
   "dependencies": {
+    "@reduxjs/toolkit": "^1.8.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.4",
     "react-router-dom": "^5.2.0",
-    "redux": "^4.1.0",
-    "redux-thunk": "^2.3.0"
+    "redux": "^4.1.0"
   }
 }
